pyproject.toml
[project]
name = "autopaper"
version = "0.1.0"
description = "CLI-driven automated newspaper generation tool with AI-powered curation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AutoPaper Contributors"}
]
maintainers = [
    {name = "AutoPaper Contributors"}
]
keywords = [
    "automation",
    "newspaper",
    "ai",
    "claude",
    "cli",
    "content-curation",
    "obsidian",
    "pdf-generation",
    "web-scraping"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "readability-lxml>=0.8.1",
    "python-dateutil>=2.8.2",
    "jinja2>=3.1.3",
    "weasyprint>=60.0",
    "pyyaml>=6.0.1",
    "anthropic>=0.18.0",
    "pillow>=10.0.0",
    "python-slugify>=8.0.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/OldCoderIsMe/AutoPaper"
Documentation = "https://github.com/OldCoderIsMe/AutoPaper#readme"
Repository = "https://github.com/OldCoderIsMe/AutoPaper"
"Bug Tracker" = "https://github.com/OldCoderIsMe/AutoPaper/issues"
Changelog = "https://github.com/OldCoderIsMe/AutoPaper/blob/main/CHANGELOG.md"

[project.scripts]
autopaper = "autopaper.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
